==> TE-WAS/Jupyter_notebook/1_liftover_te_transcript.ipynb <==
import subprocess
import sys
import os
import shutil
import pandas as pd

def shell_do(command, log=False, return_log=False):
    print(f'Executing: {(" ").join(command.split())}', file=sys.stderr)

    res=subprocess.run(command.split(), stdout=subprocess.PIPE)

    if log:
        print(res.stdout.decode('utf-8'))
    if return_log:
        return(res.stdout.decode('utf-8'))

# set paths
basedir = '/data/songy4/tes'
datadir = f'{basedir}/data_folder'
fusiondir = f'/data/songy4/twas/fusion_twas'
geno_path = f'{datadir}/qc_genotypes_tes_case'
geno_path_1 = f'{datadir}/qc_genotypes_tes_control'
#tr_list_path = f'{datadir}/transcript_list.txt'
#te_list_path = f'{datadir}/te_list.txt'

#pheno_path = f'{datadir}/expression_matrix_final.txt'
#coord_path = f'{datadir}/twas_coordinate.txt'
#covar_path = f'{datadir}/covariates_case.txt'

#open bim file 
bim_df = pd.read_csv(f"{geno_path}.bim", sep='\t', header=None,  names=('chr', 'rsid', 'kb', 'pos', 'a1', 'a2'))

#combine chr and pos columns
bim_df['MarkerName'] = str('chr') + bim_df.chr.astype(str) + ':' + bim_df.pos.astype(str)

#check the shape of meta
print("shape of bim data:", bim_df.shape)

#check the shape of meta
print("number of unique MarkerName in bim data:", bim_df['MarkerName'].nunique())
print("number of unique rsid in bim data:", bim_df['rsid'].nunique())
bim_df.head()

# Liftover genotype from hg38 to hg19 to match 1kg ref

# get chrN:start-end positions for liftover of genotype from hg38 to hg19 to match 1kG LD ref
lift_outname = f'{datadir}/geno_hg38_positions.bed'
bim_df['chr'] = 'chr' + bim_df['chr'].astype('str')
bim_df['end'] = bim_df['pos'] + 1
lift_out = bim_df[['chr', 'pos', 'end', 'rsid' ]].copy()
lift_out.to_csv(lift_outname, sep='\t', header=False, index=False)

# pull liftOver from UCSC
#!wget http://hgdownload.soe.ucsc.edu/admin/exe/linux.x86_64/liftOver -P /data/songy4/twas/liftover/
#!wget --timestamping 'ftp://hgdownload.cse.ucsc.edu/goldenPath/hg38/liftOver/hg38ToHg19.over.chain.gz' -O /data/songy4/twas/liftover/hg38ToHg19.over.chain.gz
#!chmod +x /data/songy4/twas/liftover/liftOver

# create command for liftover
liftOver = f'/data/songy4/twas/liftover/liftOver'
chainfile = f'/data/songy4/twas/liftover/hg38ToHg19.over.chain.gz'
liftover_output = f'{datadir}/geno_hg19_positions.bed'
unlifted = f'{datadir}/unlifted.bed'

liftover_cmd = f'\
{liftOver} {lift_outname} {chainfile} {liftover_output} {unlifted}'

shell_do(liftover_cmd)
liftover_cmd

# create map update file
lifted = pd.read_csv(liftover_output, sep='\t', header=None, names=['chr', 'pos', 'end', 'rsid' ])

bim_lift_merge = bim_df.merge(lifted, how='right', on='rsid')
lifted_bim = bim_lift_merge[['chr_x', 'rsid', 'kb', 'pos_y', 'a1', 'a2']].copy()
lifted_bim.columns = ['chr', 'rsid', 'kb', 'pos', 'a1', 'a2']

# use only lifted snps
lifted_bim['rsid'].to_csv(f'{geno_path}_hg19_lifted.snplist', sep='\t', header=False, index=False)

plink_extract_cmd = f'\
plink --bfile {geno_path}\
 --extract {geno_path}_hg19_lifted.snplist\
 --make-bed\
 --out {geno_path}_hg19_lifted'

shell_do(plink_extract_cmd)


# move bim with old positions to new file
!mv {geno_path}_hg19_lifted.bim {geno_path}_hg19_lifted_old_positions.bim

# write lifted bim to _hg19_lifted genotype name
lifted_bim.to_csv(f'{geno_path}_hg19_lifted.bim', sep='\t', header=False, index=False)

#check number of IDs in original .bim file
!cat {datadir}/qc_genotypes_tes_case_hg19_lifted.bim | wc -l

#look at the bim file (variance, chromosome and position, allele1 and allele2)
! head {datadir}/qc_genotypes_tes_case_hg19_lifted.bim

#check number of IDs in original .bim file
!cat {datadir}/qc_genotypes_tes_case_hg19_lifted.fam | wc -l

#look at the bim file (variance, chromosome and position, allele1 and allele2)
! head {datadir}/qc_genotypes_tes_case_hg19_lifted.fam

Control liftover

# get chrN:start-end positions for liftover of genotype from hg38 to hg19 to match 1kG LD ref
lift_outname_1 = f'{datadir}/geno_hg38_positions_control.bed'
bim = pd.read_csv(f'{geno_path_1}.bim', sep='\t', header=None)
bim.columns = ['chr', 'rsid', 'kb', 'pos', 'a1', 'a2']
bim['chr'] = 'chr' + bim['chr'].astype('str')
bim['end'] = bim['pos'] + 1
lift_out_1 = bim[['chr', 'pos', 'end', 'rsid' ]].copy()
lift_out_1.to_csv(lift_outname_1, sep='\t', header=False, index=False)

# create command for liftover
liftOver = f'/data/songy4/twas/liftover/liftOver'
chainfile = f'/data/songy4/twas/liftover/hg38ToHg19.over.chain.gz'
liftover_output_1 = f'{basedir}/geno_hg19_positions_control.bed'
unlifted_1 = f'{basedir}/unlifted_control.bed'

liftover_1_cmd = f'\
{liftOver} {lift_outname_1} {chainfile} {liftover_output_1} {unlifted_1}'

shell_do(liftover_1_cmd)
liftover_1_cmd

# create map update file
lifted_1 = pd.read_csv(liftover_output_1, sep='\t', header=None, names=['chr', 'pos', 'end', 'rsid' ])

bim_lift_merge_1 = bim.merge(lifted_1, how='right', on='rsid')
lifted_bim_1 = bim_lift_merge_1[['chr_x', 'rsid', 'kb', 'pos_y', 'a1', 'a2']].copy()
lifted_bim_1.columns = ['chr', 'rsid', 'kb', 'pos', 'a1', 'a2']

# use only lifted snps
lifted_bim_1['rsid'].to_csv(f'{geno_path_1}_hg19_lifted.snplist', sep='\t', header=False, index=False)

plink_extract_1_cmd = f'\
plink --bfile {geno_path_1}\
 --extract {geno_path_1}_hg19_lifted.snplist\
 --make-bed\
 --out {geno_path_1}_hg19_lifted'

shell_do(plink_extract_1_cmd)

# move bim with old positions to new file
!mv {geno_path_1}_hg19_lifted.bim {geno_path_1}_hg19_lifted_old_positions.bim

# write lifted bim to _hg19_lifted genotype name
lifted_bim_1.to_csv(f'{geno_path_1}_hg19_lifted.bim', sep='\t', header=False, index=False)

#check number of IDs in original .bim file
!cat {datadir}/qc_genotypes_tes_control_hg19_lifted.bim | wc -l

#look at the bim file (variance, chromosome and position, allele1 and allele2)
! head {datadir}/qc_genotypes_tes_control_hg19_lifted.bim

#check number of IDs in original .bim file
!cat {datadir}/qc_genotypes_tes_control_hg19_lifted.fam | wc -l

#look at the bim file (variance, chromosome and position, allele1 and allele2)
! head {datadir}/qc_genotypes_tes_control_hg19_lifted.fam

liftover coordinate file from hg38 to hg19   ---- IGNORE THIS AS POSITION IS ALREADY IN gr38.87

If you submit data to the browser in position format (chr#:##-##), the browser assumes this information is 1-based. If you submit data in any other format (BED (chr# ## ##) or otherwise), the browser will assume it is 0-based.
https://genome.ucsc.edu/FAQ/FAQtracks.html#tracks1

# get chrN:start-end positions for liftover of coordinate file from hg38 to hg19 to match 1kG LD ref
#coords_lift_outname = f'{datadir}/coords_hg38_positions.bed'
#coords = pd.read_csv(coord_path, sep='\t')
#coords = coords.rename(columns={'X.Chr':'chr', 'start':'pos', 'end':'end', 'ID':'id'})
#coords['chr'] = 'chr' + coords['chr'].astype('str')
#coords.to_csv(coords_lift_outname, sep='\t', header=False, index=False)

# create command for liftover
#liftOver = f'{basedir}/liftover/liftOver'
#chainfile = f'{basedir}/liftover/hg38ToHg19.over.chain.gz'
#coords_liftover_output = f'{basedir}/coords_hg19_positions.bed'
#coords_unlifted = f'{basedir}/coords_unlifted.bed'

#coords_liftover_cmd = f'\
#{liftOver} {coords_lift_outname} {chainfile} {coords_liftover_output} {coords_unlifted}'

#shell_do(coords_liftover_cmd)
#coords_liftover_cmd

#read lifted coordinate file
#lifted = pd.read_csv(coords_liftover_output, sep='\t', header=None, names=['X.Chr', 'start', 'end', 'ID' ])
#lifted

#drop chr from X.Chr
#lifted['X.Chr'] = lifted['X.Chr'].str.replace('chr', '')
#lifted

#save as text file  --> this is the new coordinate data 
#lifted.to_csv(r'./data_folder/twas_coordinate_lifted.txt', sep='\t' ,index=False)

#create gene_list as list to drop columns in exp_case 
#gene_list = lifted['ID']
#print("number of genes in gene_list:", len(gene_list))

#save gene_list as txt -- run only once for download
gene_list.to_csv(r'./data_folder/gene_list.txt', sep='\t' ,index=False)

Liftover control genotype from hg38 to hg19

# set paths for controls
control_geno_path = f'{datadir}/qc_genotypes_twas_control'
pheno_path = f'{datadir}/twas_expression_control.txt'
covar_path = f'{datadir}/covariates_control.txt'

# get chrN:start-end positions for liftover of genotype from hg38 to hg19 to match 1kG LD ref
lift_outname_cont = f'{datadir}/control_geno_hg38_positions.bed'
bim_cont = pd.read_csv(f'{geno_path}_control.bim', sep='\t', header=None)
bim_cont.columns = ['chr', 'rsid', 'kb', 'pos', 'a1', 'a2']
bim_cont['chr'] = 'chr' + bim_cont['chr'].astype('str')
bim_cont['end'] = bim_cont['pos'] + 1
lift_out_cont = bim_cont[['chr', 'pos', 'end', 'rsid' ]].copy()
lift_out_cont.to_csv(lift_outname_cont, sep='\t', header=False, index=False)

# create command for liftover
liftOver = f'{basedir}/liftover/liftOver'
chainfile = f'{basedir}/liftover/hg38ToHg19.over.chain.gz'
liftover_output_cont = f'{basedir}/control_geno_hg19_positions.bed'
unlifted_cont = f'{basedir}/control_unlifted.bed'

liftover_cont_cmd = f'\
{liftOver} {lift_outname_cont} {chainfile} {liftover_output_cont} {unlifted_cont}'

shell_do(liftover_cont_cmd)
liftover_cont_cmd

# create map update file
lifted_cont = pd.read_csv(liftover_output_cont, sep='\t', header=None, names=['chr', 'pos', 'end', 'rsid' ])

bim_cont = pd.read_csv(f'{geno_path}_control.bim', sep='\t', header=None, names=['chr', 'rsid', 'kb', 'pos', 'a1', 'a2'])
bim_lift_merge_cont = bim_cont.merge(lifted_cont, how='right', on='rsid')
lifted_bim_cont = bim_lift_merge_cont[['chr_x', 'rsid', 'kb', 'pos_y', 'a1', 'a2']].copy()
lifted_bim_cont.columns = ['chr', 'rsid', 'kb', 'pos', 'a1', 'a2']

# use only lifted snps
lifted_bim_cont['rsid'].to_csv(f'{geno_path}_control_hg19_lifted.snplist', sep='\t', header=False, index=False)

plink_extract_cont_cmd = f'\
plink --bfile {geno_path}_control\
 --extract {geno_path}_control_hg19_lifted.snplist\
 --make-bed\
 --out {geno_path}_control_hg19_lifted'

shell_do(plink_extract_cont_cmd)

# move bim with old positions to new file
!mv {geno_path}_control_hg19_lifted.bim {geno_path}_control_hg19_lifted_old_positions.bim

# write lifted bim to _hg19_lifted genotype name
lifted_bim_cont.to_csv(f'{geno_path}_control_hg19_lifted.bim', sep='\t', header=False, index=False)

#check number of IDs in original .bim file
!cat {datadir}/qc_genotypes_twas_control_hg19_lifted.bim | wc -l

#look at the bim file (variance, chromosome and position, allele1 and allele2)
! head {datadir}/qc_genotypes_twas_control_hg19_lifted.bim
