==> TE-WAS/Jupyter_notebook/2_gwas_sumstat_preprocessing_tes.ipynb <==
#upload some libraries
import os
import pandas.testing as tm
import h5py
import numpy as np
import pandas as pd
import math
import sys
import subprocess
import statsmodels.api as sm
from scipy import stats
import statsmodels.formula.api as smf
from patsy.builtins import *
import statistics
#import umap.umap_ as umap
#import umap as umap
from joblib import dump, load
import matplotlib.pyplot as plt
from matplotlib import style
import seaborn as sns
from sklearn.decomposition import PCA

! pip install --upgrade tables
import tables

# set paths
basedir = '/data/songy4/tes'
datadir = f'{basedir}/data_folder'
workdir = f'{datadir}/output'

#download the renormalized datasets ---- run it only once
#TElocal output files normalized with limma for a different timepoint, weren't adjusted for any covariates
! cp -R /data/LNG/Frank/TE/limma/baseline_GENE_TE_norm_adjusted.cntTable /{datadir}/
! cp -R /data/LNG/Frank/TE/limma/v04_GENE_TE_norm_adjusted.cntTable /{datadir}/
! cp -R /data/LNG/Frank/TE/limma/v06_GENE_TE_norm_adjusted.cntTable /{datadir}/
! cp -R /data/LNG/Frank/TE/limma/v08_GENE_TE_norm_adjusted.cntTable /{datadir}/

#TE location data
! cp -R /data/LNG/Frank/TE/hg38_rmsk_TE.gtf.locInd.locations /{datadir}/

#gene locations
! cp -R /data/LNG/Frank/TE/hg38_gene_annotation.gtf.locations /{datadir}/

#plate information (one of covariates among phenotype (= case vs control), sex, age, plate, expression pcs 1 to 5)
! cp -R /data/LNG/Frank/TE/TElocal_replicates/megaMetaTable.csv /{datadir}/

#genomic data (plink files)
! cp -R /data/LNG/Frank/TE/hits_plus_tags_AMPPD_Euros.bed /{datadir}/
! cp -R /data/LNG/Frank/TE/hits_plus_tags_AMPPD_Euros.bim /{datadir}/
! cp -R /data/LNG/Frank/TE/hits_plus_tags_AMPPD_Euros.fam /{datadir}/
! cp -R /data/LNG/Frank/TE/hits_plus_tags_AMPPD_Euros.log /{datadir}/
! cp -R /data/LNG/Frank/TE/hits_plus_tags_AMPPD_Euros.nosex /{datadir}/

#PD GWAS sum stats
! cp -R /data/LNG/Frank/meta5_sumstats_harmonized_hg19_hg38.csv /{datadir}/

#pheno
! cp -R /data/LNG/Frank/TE/DEMOG_DIAG.csv /{datadir}/

#/data/CARD/PD/summary_stats --> PD GWAS sum stats were based on files from here

GWAS summary stats

#open pheno file megaMetaTable.csv
statsum = pd.read_csv(f"{datadir}/meta5_sumstats_harmonized_hg19_hg38.csv", index_col=False )

print('shape of GWAS summary stats data:', statsum.shape)
statsum

#flip Allele1 and Allele2 if Freq1 is bigger than .5 and change Freq1 to negative
stat_sum = statsum.copy()
cond = (stat_sum['FREQ'] > 0.5)
stat_sum.loc[cond, ['A1', 'A2']] = stat_sum.loc[cond, ['A2', 'A1']].values
#stat_sum_flip.Allele1, stat_sum_flip.Allele2 = np.where(stat_sum_flip.Freq1 > 0.5, [stat_sum_flip.Allele2, stat_sum_flip.Allele1], [stat_sum_flip.Allele1, stat_sum_flip.Allele2])

#change Freq1
stat_sum.loc[cond, ['FREQ']] = 1-stat_sum['FREQ']

#change Effect value
stat_sum.loc[cond, 'BETA'] *= -1

stat_sum
##no need to change  StdErr, P-value

#drop two rows Freq1 is smaller than 0.01
stat_sum.drop(stat_sum[stat_sum.FREQ < 0.01].index, inplace=True)
stat_sum = stat_sum.reset_index(drop=True)
#stat_sum = stat_sum.drop(['index'], axis=1)

#drop unwanted columns
stat_sum = stat_sum[['RSID', 'REF', 'ALT', 'BETA', 'SE', 'P','BP_HG19']]

#check the unique number of MarkerName in meta
print("Unique RSID in meta data:", stat_sum['RSID'].nunique())

stat_sum

#print duplicated rows
stat_sum_dup = stat_sum[stat_sum['RSID'].duplicated(keep=False)]
print(stat_sum_dup.head())

#count how many times patient number is duplicated
duplicate = stat_sum_dup.groupby(stat_sum_dup['RSID'].tolist(),as_index=False).size()
print(duplicate)

#drop rows with RSID as .
stat_sum = stat_sum[stat_sum.RSID != '.']

#drop duplicated ID into 1 
stat_sum = stat_sum.drop_duplicates(['RSID']).reset_index(drop=True)

#calculate z score with effect size and standard error (z = (score-mean)/se)
def z_score(a, b):
    if b == 0:
        return 0.0
    return float(a)/b

stat_sum['Z'] = np.vectorize(z_score)(stat_sum['BETA'], stat_sum['SE'])
stat_sum

#calculate chisq
stat_sum['chisq'] = stat_sum['Z']**2

#calculate lambda (median(z^2) / 0.455)
lambda_statsum = ((stat_sum['P'].median())/0.456)
print(lambda_statsum)
stat_sum

#rename columns 
sumstat = stat_sum.copy()
sumstat.rename(columns={'RSID':'SNP', 'REF':'A1', 'ALT':'A2'}, inplace=True)

#drop BETA, SE, P, BP_HG19, chisq colunms 
sumstat = sumstat.drop(['BETA', 'SE', 'P', 'BP_HG19', 'chisq'], axis=1)

sumstat = sumstat.reset_index(drop=True)
print("shape of meta data:", sumstat.shape)
sumstat.head()

#save meta_df as txt -- run only once for download
sumstat.to_csv(r'./data_folder/meta.txt', sep='\t' ,index=False)
stat_sum.to_csv(r'./data_folder/meta_detail.txt', sep='\t' ,index=False)
