==> TE-WAS/Jupyter_notebook/3_calculate_weights_case_te_transcript.ipynb <==
#clone fusion_twas library before running this
!git clone https://github.com/gusevlab/fusion_twas.git

import subprocess
import sys
import os
import shutil
import pandas as pd
import numpy as np

def shell_do(command, log=False, return_log=False):
    print(f'Executing: {(" ").join(command.split())}', file=sys.stderr)

    res=subprocess.run(command.split(), stdout=subprocess.PIPE)

    if log:
        print(res.stdout.decode('utf-8'))
    if return_log:
        return(res.stdout.decode('utf-8'))

#set paths
basedir = '/data/songy4/tes'
datadir = f'{basedir}/data_folder'
fusiondir = f'/data/songy4/twas/fusion_twas'
geno_path_case = f'{datadir}/qc_genotypes_tes_case_hg19_lifted'
geno_path_cont = f'{datadir}/qc_genotypes_tes_control_hg19_lifted'
transcript_list_path = f'{datadir}/transcript_list.txt'
te_list_path = f'{datadir}/te_list.txt'
pheno_path_tecase = f'{datadir}/baseline_te_case.txt'
pheno_path_tecont = f'{datadir}/baseline_te_control.txt'
pheno_path_trcase = f'{datadir}/base_transcript_case.txt'
pheno_path_trcont = f'{datadir}/base_transcript_control.txt'
coord_path_te = f'{datadir}/te_coordinate.txt'
coord_path_tr = f'{datadir}/transcript_coordinate.txt'
covar_path_case = f'{datadir}/te_covariates_case.txt'
covar_path_cont = f'{datadir}/te_covariates_control.txt'
gcta = f'{fusiondir}/gcta_nr_robust'
gemma = f'gemma-0.98.3-linux-static'
fusion_ldref_basename = f'{fusiondir}/LDREF/1000G.EUR'
fusion_compute_weights_script = f'{fusiondir}/FUSION.compute_weights.R'

!mkdir --parents {basedir}/output/weights_case
!mkdir --parents {basedir}/output/tmp_case


# Pipeline for TWAS run

# get gene list
##test with couple genes first before running all gene_list
#transcript_list_df = pd.read_csv(transcript_list_path, engine='c',sep='\t')
#transcript_list = list(transcript_list_df.ID)
#transcript_list = ['ENST00000000233.10', 'ENST00000000412.8', 'ENST00000000442.11'] # for testing
# gene_list = ['ENSG00000186092']
#pheno_trcase = pd.read_csv(pheno_path_trcase, engine='c', sep='\t')
#pheno_trcont = pd.read_csv(pheno_path_trcont, engine='c',sep='\t')
#coords = pd.read_csv(coord_path_tr, engine='c',sep='\t')
#covars_case = pd.read_csv(covar_path_case, engine='c',sep='\t')
#covars_cont = pd.read_csv(covar_path_cont, sep='\t')

# now put together pipeline
covars = pd.read_csv(covar_path_case, sep='\t')

for i in range(21 ,23): 
    pheno_i = pd.read_csv(f'{datadir}/base_transcript_case_chr{i}.txt', engine='c')
    coords_i = pd.read_csv(f'{datadir}/transcript_coordinate_chr{i}.txt',  engine='c')
    gene_list_df_i = pd.read_csv(f'{datadir}/transcript_list_chr{i}.txt', engine='c')
    gene_list_i = list(gene_list_df_i.ID)
    
    compweights_swarmfile_i = f'{basedir}/compute_weights_{i}.swarm'

    with open(compweights_swarmfile_i, 'w') as f:
    
        for gene in gene_list_i:
            OUT = f'output/tmp_case/{gene}'
            FINAL_OUT = f'output/weights_case/{gene}'
            #get chr start stop
            _chr = coords_i.loc[coords_i.ID == gene, 'X.Chr'].item()
            _start = coords_i.loc[coords_i.ID == gene, 'start'].item()-0.5e6
            _stop = coords_i.loc[coords_i.ID == gene, 'end'].item()+0.5e6

            if _start < 0:
                _start = 0

            _start = int(_start)
            _stop = int(_stop)

            pheno_i[['FID','IID', gene]].to_csv(f'{OUT}.pheno', sep='\t', header=False, index=False)
    #         pheno[['FID','IID', gene]].to_csv(_phenoname, sep='\t', header=False, index=False)

            plink_cmd_i = f'\
plink --bfile {geno_path_case} \
--pheno {OUT}.pheno \
--keep {OUT}.pheno \
--chr {_chr} \
--from-bp {_start} \
--to-bp {_stop} \
--extract {fusion_ldref_basename}.{_chr}.bim \
--make-bed \
--out {OUT}'  

#        shell_do(plink_cmd, log=True, return_log=True)

            fusion_cmd_i = f'\
Rscript {fusion_compute_weights_script} \
--bfile {OUT} \
--tmp {OUT}.tmp \
--out {FINAL_OUT} \
--PATH_gemma {gemma} \
--PATH_plink plink \
--PATH_gcta {gcta} \
--covar {covar_path_case} \
--verbose 2 \
--save_hsq \
--models top1,lasso,enet'
##include "blup,bslmm" to the --models in fusion_cmd for future     
# 
#        shell_do(fusion_cmd, log=True, return_log=True)

            f.write(f'{plink_cmd_i} && {fusion_cmd_i}\n')
        f.close()

# run swarm
#compweights_swarmfile_3 = f'{basedir}/compute_weights_3.swarm'
#swarm_cmd_3 = f'swarm -f {compweights_swarmfile_3} -g 16 -t 16 --time=10:00:00 --logdir swarm --gres=lscratch:20 --module plink,GEMMA/0.96 --partition=norm'
# shell_do(swarm_cmd)
#print(swarm_cmd_3)
#!{swarm_cmd_3}
##check swarm jobs state in hpc.nih.gov/dashboard/
##check .e, .o, and .log files of swarm jobs in swarm folder (same name as swarm jobid in hpc dashboard)

# run swarm
for i in range(21,23): 
    compweights_swarmfile_i = f'{basedir}/compute_weights_{i}.swarm'
    swarm_cmd_i = f'swarm -f {compweights_swarmfile_i} -g 8 -t 2 --time=10:00:00 --logdir swarm --gres=lscratch:20 --module plink,GEMMA/0.96 --partition=norm'
#--devel
# shell_do(swarm_cmd)
    print(swarm_cmd_i)
    !{swarm_cmd_i}
###for this specific project, 8GB memory and 2 CPU are enough###
##check swarm jobs state in hpc.nih.gov/dashboard/
##check .e, .o, and .log files of swarm jobs in swarm folder (same name as swarm jobid in hpc dashboard)

#check number of weights in output/weights folder 
###run this in terminal

#go to the weights folder
!cd /data/songy4/tes/output/weights_case
#check how many files are in total
! ls | wc -l
#there are 218193
#check how many .RDat file is in weights folder
!ls -lR ./*.RDat | wc -l
#there are 9482 RDat files

TWAS

#open gene_chr_start_stop.tab -- this has chr and position info
chr_pos_df = pd.read_csv(coord_path_tr, sep='\t')
#rearrange the columns 
chr_pos_df = chr_pos_df[['ID', 'X.Chr','start','end']]
#change column names in chr_pos_df
chr_pos_df.columns = ['ID', 'CHR', 'P0', 'P1']

chr_pos_df

#make gene_path.csv
! ls -l ./output/weights_case | grep -E 'ENSG|.RDat' | awk '{print $9", ./weights/"$9}' | sed 's/.wgt.RDat,/,/g' > ./data_folder/gene_path_case.csv

## list files in weights folder, search lines with same pattern as 'ENSG'
## scan file line by line, splits each input line
## substitute .wgt.RDat, with , in all occurrences (/g means global replacement)

#import gene_path.csv 
path = pd.read_csv(f"{datadir}/gene_path_case.csv", names=['ID', 'WGT'])
path.head

#remove .hsq rows --> didn't need in this
#path = path[~path.WGT.str.contains(".hsq")]
#path.head

#save path as csv -- run only once for download
#path.to_csv(r'./data_folder/gene_path_case.csv' ,index=False)

#merge chr_pos_df and path
merged = chr_pos_df.merge(path, how='inner', on='ID')
merged.head

#check chr 
chr_list = merged['CHR'].unique().tolist()
chr_list

#drop ./weights/ from WGT
merged['WGT'] = merged['WGT'].str.replace('./weights/', '')
merged.head

#rearrange columns and save the file as synapse.pos
output = merged[['WGT', 'ID', 'CHR', 'P1', 'P0']]
output.to_csv(r'./data_folder/synapse_case.pos', index=False, sep="\t")

#check the RDat file sizes - check few files from above in terminal: the file sizes are ranging 500-1200
#ls -l filename   #Displays Size of the specified file
#ls -l *          #Displays Size of All the files in the current directory
#ls -al *         #Displays Size of All the files including hidden files in the current directory
#ls -al dir/      #Displays Size of All the files including hidden files in the 'dir' directory

!ls -l ENST00000426108.1.wgt.RDat 

#set paths
sumstat_path = f'{datadir}/meta.txt'
weights_path = f'{datadir}/synapse_case.pos'
weights_dir = f'{basedir}/output/weights_case/'
fusiondir = f'/data/songy4/twas/fusion_twas'
fusion_ldref_basename = f'{fusiondir}/LDREF/1000G.EUR'
fusion_assoc_script = f'{fusiondir}/FUSION.assoc_test.R'

!mkdir --parents output/pd_case

sumstat = pd.read_csv(sumstat_path, sep='\t')

##check if weights_dir should be f'{basedir}/output/weights'####

#fusion calculation -- take about 3 min each chromosome 
for i in range(1, 23):
    PD_OUT = f'output/pd_case/PD_case.{i}.dat'
    fusion_cmd_i = f'\
    Rscript {fusion_assoc_script} \
    --sumstats {sumstat_path} \
    --weights {weights_path} \
    --weights_dir {weights_dir} \
    --ref_ld_chr {fusion_ldref_basename}. \
    --chr {i} \
    --out {PD_OUT}'
    shell_do(fusion_cmd_i)

# running list of questions
1. what about non-autosome transcripts? i.e. MT
2. how many snps in common between ref and geno?
3. how many snps per transcript should we expect (0-500?)


p_in = 'tmp/ENSG00000188976.tmp.cv'
f"gemma -miss 1 -maf 0 -r2 1 -rpace 1000 -wpace 1000 -bfile {p_in} -bslmm 2 -o ENSG00000188976"
